requirements.txt
APScheduler==3.6.3
blueprint==3.4.2
click==7.1.2
colorhash==1.0.2
configparser==5.0.1
dnspython==2.0.0
Faker==4.14.2
Flask==1.1.2
Flask-MonitoringDashboard==3.1.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.19.4
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
psutil==5.7.3
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
tzlocal==2.1
Werkzeug==1.0.1
